Separate the UI and API server
upto this point, UI and back-end server were on the same port.
It was ok for small development application. But when application
becomes large, it is not ideal to put UI and server on the same port.
So, I updated the architecture, and separate the UI and back-end
API(GraphQL) server and run them on two different port.